completed lab03 debugging part 2

+++ lab03/src/adventure/SpeciesListStage.java
@@ -105,6 +105,6 @@ public class SpeciesListStage implements AdventureStage {
                 copy.remove(o);
             }
         }
-        return similarObjects / listOne.size();
+        return (double) similarObjects / listOne.size();
     }
 }
